added win screen and change title screen a little bit
I changed the options button to be an exit button instead (which actually works when you boot it up properly like .exe).

And yeah I added a win screen as well that has a next level button that just has the method to play the convenience level attached to it.

and added music to everything + changed the Start function in the audio manager to have it play a different theme depending on what scene is currently loaded.